feat: migrate FormInputFile.spec
