--- cypress/component/Forms/FormInputFile.spec.tsx
@@ -1,112 +1,0 @@
-import React from 'react'
-import '../../../src/components/forms/formComponents.css'
-
-// Simple test component that replicates FormInputFile behavior
-const TestFormInputFile = ({
-  id,
-  disabled = false,
-  uploadText = 'Upload a file',
-  validation = { valid: true },
-}: {
-  id: string
-  disabled?: boolean
-  uploadText?: string
-  validation?: { valid: boolean }
-}) => {
-  return (
-    <div style={{ display: 'flex', flexDirection: 'row', alignItems: 'center' }}>
-      <div className={`form-file-upload ${disabled ? 'disabled' : ''}`}>
-        <input
-          id={id}
-          type="file"
-          style={{ display: 'none' }}
-          disabled={disabled}
-        />
-        <label
-          {...(!disabled && { htmlFor: id })}
-          className={`form-file-label ${!validation.valid ? 'errored' : ''}`}
-        >
-          <span>📁</span>
-          {' '}
-          {/* Using emoji instead of MUI icon to avoid import issues */}
-          {uploadText}
-        </label>
-      </div>
-    </div>
-  )
-}
-
-describe('FormInputFile Component', () => {
-  it('should render upload button when not disabled', () => {
-    cy.mount(<TestFormInputFile id="test-file-input" />)
-
-    cy.get('.form-file-upload').should('exist')
-    cy.get('.form-file-upload').should('not.have.class', 'disabled')
-    cy.get('label.form-file-label').should('have.attr', 'for', 'test-file-input')
-    cy.get('input[type="file"]').should('exist')
-    cy.get('input[type="file"]').should('not.be.disabled')
-  })
-
-  it('should disable upload button when disabled prop is true', () => {
-    cy.mount(<TestFormInputFile id="test-file-input" disabled={true} />)
-
-    cy.get('.form-file-upload').should('exist')
-    cy.get('.form-file-upload').should('have.class', 'disabled')
-    cy.get('label.form-file-label').should('not.have.attr', 'for')
-    cy.get('input[type="file"]').should('exist')
-    cy.get('input[type="file"]').should('be.disabled')
-  })
-
-  it('should show custom upload text', () => {
-    cy.mount(<TestFormInputFile id="test-file-input" uploadText="Choose your file" />)
-
-    cy.get('label.form-file-label').should('contain.text', 'Choose your file')
-  })
-
-  it('should appear disabled in read-only DAR and not trigger file selection', () => {
-    cy.mount(<TestFormInputFile id="test-file-input" disabled={true} />)
-
-    // Verify visual disabled state
-    cy.get('.form-file-upload').should('have.class', 'disabled')
-    // Note: CSS opacity might be computed differently, so let's just check it's less than 1
-    cy.get('.form-file-upload').should('satisfy', ($el) => {
-      const opacity = window.getComputedStyle($el[0]).opacity
-      return parseFloat(opacity) < 1
-    })
-
-    // Verify label doesn't have htmlFor attribute (prevents file dialog)
-    cy.get('label.form-file-label').should('not.have.attr', 'for')
-
-    // Verify input is disabled
-    cy.get('input[type="file"]').should('be.disabled')
-
-    // Click on the label - it should not associate with the file input
-    cy.get('label.form-file-label').click()
-
-    // Since the label doesn't have a 'for' attribute when disabled,
-    // clicking it won't trigger the file input
-    cy.get('input[type="file"]').should('not.have.focus')
-  })
-
-  it('should have proper cursor styling when disabled', () => {
-    cy.mount(<TestFormInputFile id="test-file-input" disabled={true} />)
-
-    // Verify disabled cursor styling is applied via CSS
-    // Note: cursor styles might not be testable in headless mode, so let's verify the CSS class exists
-    cy.get('.form-file-upload').should('have.class', 'disabled')
-    cy.get('.form-file-upload.disabled').should('exist')
-  })
-
-  it('should enable file selection when clicking enabled upload button', () => {
-    cy.mount(<TestFormInputFile id="test-file-input" disabled={false} />)
-
-    // Verify enabled state
-    cy.get('.form-file-upload').should('not.have.class', 'disabled')
-    cy.get('label.form-file-label').should('have.attr', 'for', 'test-file-input')
-    cy.get('input[type="file"]').should('not.be.disabled')
-
-    // The label should be properly linked to the input when enabled
-    cy.get('label.form-file-label').click()
-    // We can't test file dialog opening directly, but we can verify the association exists
-  })
-})
